pyproject.toml
[project]
name = "proposify-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.5.3",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "dj-rest-auth[with-social]>=7.0.1",
    "django==5.2.7",
    "django-cors-headers>=4.9.0",
    "django-countries>=7.6.1",
    "django-filter>=25.2",
    "django-phonenumber-field[phonenumberslite]>=8.3.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "environs>=14.3.0",
    "huggingface-hub>=1.1.2",
    "pillow>=12.0.0",
    "stripe>=13.1.1",
    "uvicorn>=0.38.0",
    "whitenoise>=6.11.0",
]
